+++ front-end/my-app/package.json
@@ -18,6 +18,7 @@
     "pokenode-ts": "^1.16.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "semantic-ui-css": "^2.4.1",
     "semantic-ui-react": "^2.1.3",
