+++ front-end/my-app/src/App.css
@@ -12,6 +12,7 @@ h1 {
   margin-bottom: -2px;
 }
 
+
 .image {
   left: 0;
   right: 0;
@@ -32,3 +33,4 @@ h1 {
   display:flex;
   flex-direction: column;
 }
+
